Move the EthereumClient to the place where it should be.
