+++ linera-ethereum/src/provider.rs
@@ -1,31 +1,13 @@
 // Copyright (c) Zefchain Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use alloy_primitives::{Address, Bytes, U256};
-use alloy::{
-    network::{Ethereum, EthereumWallet},
-    providers::{
-        fillers::{BlobGasFiller, ChainIdFiller, FillProvider, GasFiller, JoinFill, NonceFiller, WalletFiller},
-        Provider, ProviderBuilder, RootProvider,
-    },
-    rpc::types::eth::{
-        request::{TransactionInput, TransactionRequest},
-        Filter,
-    },
-    signers::local::PrivateKeySigner,
-    transports::http::reqwest::{header::CONTENT_TYPE, Client},
-};
+use alloy::transports::http::reqwest::{header::CONTENT_TYPE, Client};
 use async_lock::Mutex;
 use async_trait::async_trait;
-use url::Url;
-
-use crate::client::{EthereumQueries, JsonRpcClient};
-
-pub type HttpProvider = RootProvider<Ethereum>;
 
 use crate::{
-    client::get_block_id,
-    common::{event_name_from_expanded, parse_log, EthereumEvent, EthereumServiceError},
+    client::JsonRpcClient,
+    common::EthereumServiceError,
 };
 
 /// The Ethereum endpoint and its provider used for accessing the Ethereum node.
@@ -62,102 +44,5 @@ impl EthereumClientSimplified {
     pub fn new(url: String) -> Self {
         let id = Mutex::new(1);
         Self { url, id }
-    }
-}
-
-#[derive(Clone)]
-pub struct EthereumClient {
-    pub provider: FillProvider<JoinFill<JoinFill<alloy::providers::Identity, JoinFill<GasFiller, JoinFill<BlobGasFiller, JoinFill<NonceFiller, ChainIdFiller>>>>, WalletFiller<EthereumWallet>>, RootProvider<Ethereum>>,
-}
-
-#[async_trait]
-impl EthereumQueries for EthereumClient {
-    type Error = EthereumServiceError;
-
-    async fn get_accounts(&self) -> Result<Vec<String>, EthereumServiceError> {
-        Ok(self
-            .provider
-            .get_accounts()
-            .await?
-            .into_iter()
-            .map(|x| format!("{:?}", x))
-            .collect::<Vec<_>>())
-    }
-
-    async fn get_block_number(&self) -> Result<u64, EthereumServiceError> {
-        Ok(self.provider.get_block_number().await?)
-    }
-
-    async fn get_balance(
-        &self,
-        address: &str,
-        block_number: u64,
-    ) -> Result<U256, EthereumServiceError> {
-        let address = address.parse::<Address>()?;
-        let block_id = get_block_id(block_number);
-        let request = self.provider.get_balance(address).block_id(block_id);
-        Ok(request.await?)
-    }
-
-    async fn read_events(
-        &self,
-        contract_address: &str,
-        event_name_expanded: &str,
-        from_block: u64,
-        to_block: u64,
-    ) -> Result<Vec<EthereumEvent>, EthereumServiceError> {
-        let contract_address = contract_address.parse::<Address>()?;
-        let event_name = event_name_from_expanded(event_name_expanded);
-        let filter = Filter::new()
-            .address(contract_address)
-            .event(&event_name)
-            .from_block(from_block)
-            .to_block(to_block - 1);
-        let events = self.provider.get_logs(&filter).await?;
-        events
-            .into_iter()
-            .map(|x| parse_log(event_name_expanded, x))
-            .collect::<Result<_, _>>()
-    }
-
-    async fn non_executive_call(
-        &self,
-        contract_address: &str,
-        data: Bytes,
-        from: &str,
-        block: u64,
-    ) -> Result<Bytes, EthereumServiceError> {
-        let contract_address = contract_address.parse::<Address>()?;
-        let from = from.parse::<Address>()?;
-        let input = TransactionInput::new(data);
-        let tx = TransactionRequest::default()
-            .from(from)
-            .to(contract_address)
-            .input(input);
-        let block_id = get_block_id(block);
-        let eth_call = self.provider.call(tx).block(block_id);
-        Ok(eth_call.await?)
-    }
-}
-
-impl EthereumClient {
-    /// Connects to an existing Ethereum node and creates an `EthereumClient`
-    /// if successful.
-    pub fn new(url: String) -> Result<Self, EthereumServiceError> {
-        println!("EthereumClient, new, step 1");
-        let rpc_url = Url::parse(&url)?;
-        println!("EthereumClient, new, step 2");
-        let pk: PrivateKeySigner =
-            "0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80".parse().unwrap();
-        println!("EthereumClient, new, step 3");
-        let wallet = EthereumWallet::from(pk);
-        println!("EthereumClient, new, step 4");
-        let provider = ProviderBuilder::new()
-            .wallet(wallet.clone())
-            .connect_http(rpc_url);
-        println!("EthereumClient, new, step 5");
-        let endpoint = Self { provider };
-        println!("EthereumClient, new, step 6");
-        Ok(endpoint)
     }
 }
